Revert "Moved Result to exceptions.handling"
